Compiled all source files into a VS Project
In order to better test code, I have taken all of the code committed to
this repo into a single VS Project file and have cleaned up some of the
code to be more in line with the Code Complete layout and style
suggestions. In addition, I have updated (but not completely
implemented) the create and drop functions for the Database.