Update relation and grammar files
Updated my insert_into for both cases of input, corrected some
input/output of grammar. Added the is_header function to the relation
class to check if the inputted index of the header vector is a key.
Added the key_exists function to check if the inputted tuple already
exists in the map as a key. Added the constructor for the table. Added a
get_table function to return a copy of the table. Added a show function
to print out the contents of the map to the console.